+++ src/main/kotlin/icu/windea/pls/lang/codeInsight/documentation/CwtDocumentationManager.kt
@@ -331,7 +331,7 @@ object CwtDocumentationManager {
             val paths = ParadoxModifierManager.getModifierIconPaths(name, contextElement)
             paths.firstNotNullOfOrNull { path ->
                 val iconSelector = ParadoxFilePathSearch.selector(project, referenceElement).contextSensitive()
-                ParadoxFilePathSearch.searchIcon(path, iconSelector).find()
+                ParadoxFilePathSearch.searchModifierIcon(path, iconSelector).find()
             }
         }
         // 如果没找到的话，不要在文档中显示相关信息