+++ src/main/kotlin/icu/windea/pls/lang/search/ParadoxFilePathSearch.kt
@@ -6,7 +6,6 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.search.searches.ExtensibleQueryFactory
 import com.intellij.util.QueryExecutor
 import icu.windea.pls.config.configExpression.CwtDataExpression
-import icu.windea.pls.config.configExpression.CwtDataExpressionRole
 import icu.windea.pls.lang.fileInfo
 import icu.windea.pls.lang.search.searchers.ParadoxFilePathSearcher
 import icu.windea.pls.lang.search.util.ParadoxSearchParameters
@@ -31,14 +30,18 @@ class ParadoxFilePathSearch : ExtensibleQueryFactory<VirtualFile, ParadoxFilePat
      *
      * @param filePath 相对于入口目录的文件路径。或者写在脚本文件中的路径引用表达式。
      * @param configExpression 对应的规则表达式。拥有数种写法的文件路径表达式。
-     * @param ignoreLocale 使用指定的完整文件路径进行查找本地化文件时，是否忽略文件名中的本地化语言环境。
      * @param selector 查询选择器。用于指定如何选择需要查找的文件，尤其是当存在覆盖与重载的情况时。
+     * @param ignoreCase 查找文件时，是否忽略文件名的大小写。仅适用于指定了 [filePath] 且未指定 [configExpression] 的场合。
+     * @param ignoreExtension 查找文件时，是否忽略文件名中的扩展名。仅适用于指定了 [filePath] 且未指定 [configExpression] 的场合。这意味着 [filePath] 的格式应形如 `some/icon`，而非 `some/icon.dds`。
+     * @param ignoreLocale 查找文件时，是否忽略文件名中的本地化语言环境。仅适用于指定了 [filePath] 且未指定 [configExpression] 的场合。
      */
     data class Parameters(
         val filePath: String?,
         val configExpression: CwtDataExpression?,
-        val ignoreLocale: Boolean,
         override val selector: Selector,
+        val ignoreCase: Boolean = false,
+        val ignoreExtension: Boolean = false,
+        val ignoreLocale: Boolean = false,
     ) : ParadoxSearchParameters<VirtualFile>
 
     class Selector(project: Project, context: Any?) : ParadoxSearchSelector<VirtualFile>(project, context) {
@@ -46,8 +49,6 @@ class ParadoxFilePathSearch : ExtensibleQueryFactory<VirtualFile, ParadoxFilePat
     }
 
     companion object {
-        private val iconExpression = CwtDataExpression.resolve("icon[]", CwtDataExpressionRole.Value)
-
         @JvmField val EP_NAME = ExtensionPointName<QueryExecutor<VirtualFile, Parameters>>("icu.windea.pls.search.filePathSearch")
         @JvmField val INSTANCE = ParadoxFilePathSearch()
 
@@ -57,8 +58,8 @@ class ParadoxFilePathSearch : ExtensibleQueryFactory<VirtualFile, ParadoxFilePat
 
         /** @see Parameters */
         @JvmStatic
-        fun search(filePath: String?, configExpression: CwtDataExpression? = null, selector: Selector, ignoreLocale: Boolean = false): ParadoxUnaryQuery<VirtualFile> {
-            return INSTANCE.createParadoxQuery(Parameters(filePath, configExpression, ignoreLocale, selector))
+        fun search(filePath: String?, configExpression: CwtDataExpression? = null, selector: Selector, ignoreCase: Boolean = false, ignoreExtension: Boolean = false, ignoreLocale: Boolean = false): ParadoxUnaryQuery<VirtualFile> {
+            return INSTANCE.createParadoxQuery(Parameters(filePath, configExpression, selector, ignoreCase, ignoreExtension, ignoreLocale))
         }
 
         /**
@@ -66,9 +67,9 @@ class ParadoxFilePathSearch : ExtensibleQueryFactory<VirtualFile, ParadoxFilePat
          * @see ChronicleConstants.imageFileExtensions
          */
         @JvmStatic
-        fun searchImage(filePath: String?, configExpression: CwtDataExpression? = null, selector: Selector, ignoreLocale: Boolean = false): ParadoxUnaryQuery<VirtualFile> {
+        fun searchImage(filePath: String?, configExpression: CwtDataExpression? = null, selector: Selector, ignoreCase: Boolean = false, ignoreExtension: Boolean = false): ParadoxUnaryQuery<VirtualFile> {
             val selector = selector.withFileExtensions(*ChronicleConstants.imageFileExtensions) // 3.0.1 optimize: limit file extensions
-            return INSTANCE.createParadoxQuery(Parameters(filePath, configExpression, ignoreLocale, selector))
+            return search(filePath, configExpression, selector, ignoreCase, ignoreExtension)
         }
 
         /**
@@ -76,10 +77,9 @@ class ParadoxFilePathSearch : ExtensibleQueryFactory<VirtualFile, ParadoxFilePat
          * @see ChronicleConstants.imageFileExtensions
          */
         @JvmStatic
-        fun searchIcon(filePath: String?, selector: Selector, ignoreLocale: Boolean = false): ParadoxUnaryQuery<VirtualFile> {
-            val configExpression = iconExpression
+        fun searchModifierIcon(filePath: String?, selector: Selector): ParadoxUnaryQuery<VirtualFile> {
             val selector = selector.withFileExtensions(*ChronicleConstants.imageFileExtensions) // 3.0.1 optimize: limit file extensions
-            return search(filePath, configExpression, selector, ignoreLocale)
+            return search(filePath, null, selector, ignoreCase = true, ignoreExtension = true)
         }
 
         /**