+++ src/main/kotlin/icu/windea/pls/lang/search/searchers/ParadoxFilePathSearcher.kt
@@ -12,9 +12,11 @@ import com.intellij.util.indexing.FileBasedIndex
 import icu.windea.pls.ChronicleFacade
 import icu.windea.pls.config.configExpression.CwtDataExpression
 import icu.windea.pls.core.annotations.Optimized
+import icu.windea.pls.core.collections.CaseInsensitiveStringSet
 import icu.windea.pls.core.collections.forEachFast
 import icu.windea.pls.core.collections.mapFast
 import icu.windea.pls.core.collections.process
+import icu.windea.pls.core.convertPath
 import icu.windea.pls.core.toPsiFile
 import icu.windea.pls.ep.resolve.expression.ParadoxPathReferenceExpressionSupport
 import icu.windea.pls.lang.analysis.ParadoxAnalysisManager
@@ -32,6 +34,8 @@ import icu.windea.pls.model.ParadoxGameType
  */
 @Optimized
 class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSearch.Parameters>() {
+    // com.intellij.psi.search.FilenameIndex
+
     override fun processQuery(queryParameters: ParadoxFilePathSearch.Parameters, consumer: Processor<in VirtualFile>) {
         ProgressManager.checkCanceled()
         val context = queryParameters.createContext()
@@ -40,40 +44,17 @@ class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSe
 
     private fun processQuery(context: Context, consumer: Processor<in VirtualFile>): Boolean {
         if (!context.isValid()) return true
-        val indexId = ChronicleIndexKeys.FilePath
         if (context.configExpression == null) {
             if (context.filePath == null) {
-                val keys = FileBasedIndex.getInstance().getAllKeys(indexId, context.project)
-                return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
-                    ProgressManager.checkCanceled()
-                    ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
-                    if (!matchesGameType(context, file)) return@p true // check game type at file level
-                    consumer.process(file)
-                }
+                return processAllFilePaths(context, consumer)
             } else {
                 if (context.filePath.isEmpty()) return true
-                val keys = getFilePaths(context, context.filePath)
-                return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
-                    ProgressManager.checkCanceled()
-                    ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
-                    if (!matchesGameType(context, file)) return@p true // check game type at file level
-                    consumer.process(file)
-                }
+                return processFilePath(context, consumer)
             }
         } else {
             val support = ParadoxPathReferenceExpressionSupport.get(context.configExpression.type) ?: return true
             if (context.filePath == null) {
-                val keys = mutableSetOf<String>()
-                FileBasedIndex.getInstance().processAllKeys(indexId, p@{ p ->
-                    if (!support.matches(context.configExpression, context.contextElement, p)) return@p true
-                    keys.add(p)
-                }, context.scope, null)
-                return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
-                    ProgressManager.checkCanceled()
-                    ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
-                    if (!matchesGameType(context, file)) return@p true // check game type at file level
-                    consumer.process(file)
-                }
+                return processFilePathWithConfigExpression(context, support, consumer)
             } else {
                 if (context.filePath.isEmpty()) return true
                 val canResolve = support.canResolve(context.configExpression, context.filePath)
@@ -87,7 +68,65 @@ class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSe
         }
     }
 
+    private fun processAllFilePaths(context: Context, consumer: Processor<in VirtualFile>): Boolean {
+        val indexId = ChronicleIndexKeys.FilePath
+        val keys = FileBasedIndex.getInstance().getAllKeys(indexId, context.project)
+        return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
+            ProgressManager.checkCanceled()
+            ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
+            if (!matchesGameType(context, file)) return@p true // check game type at file level
+            consumer.process(file)
+        }
+    }
+
+    private fun processFilePath(context: Context, consumer: Processor<in VirtualFile>): Boolean {
+        // 3.0.2 can be case-insensitive and/or locale-insensitive here
+        val filePath = context.filePath ?: return true
+        val indexId = ChronicleIndexKeys.FilePath
+        val filePaths = getFilePaths(context, filePath)
+        if (filePaths.isEmpty()) return true
+        val keys = mutableSetOf<String>()
+        if (context.ignoreCase || context.ignoreExtension) {
+            val expected = if (context.ignoreCase) CaseInsensitiveStringSet().apply { addAll(filePaths) } else filePaths
+            FileBasedIndex.getInstance().processAllKeys(indexId, p@{ p ->
+                val actual = if (context.ignoreExtension) p.convertPath { b, _ -> b } else p
+                if (!expected.contains(actual)) return@p true
+                keys.add(p)
+                true // continue processing
+            }, context.scope, null)
+        } else {
+            keys.addAll(filePaths)
+        }
+        if (keys.isEmpty()) return true
+        return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
+            ProgressManager.checkCanceled()
+            ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
+            if (!matchesGameType(context, file)) return@p true // check game type at file level
+            consumer.process(file)
+        }
+    }
+
+    private fun processFilePathWithConfigExpression(context: Context, support: ParadoxPathReferenceExpressionSupport, consumer: Processor<in VirtualFile>): Boolean {
+        // 3.0.2 should be case-sensitive here, should be enough
+        val configExpression = context.configExpression ?: return true
+        val indexId = ChronicleIndexKeys.FilePath
+        val keys = mutableSetOf<String>()
+        FileBasedIndex.getInstance().processAllKeys(indexId, p@{ p ->
+            if (!support.matches(configExpression, context.contextElement, p)) return@p true
+            keys.add(p)
+            true // continue processing
+        }, context.scope, null)
+        if (keys.isEmpty()) return true
+        return FileBasedIndex.getInstance().processFilesContainingAnyKey(indexId, keys, context.scope, null, null) p@{ file ->
+            ProgressManager.checkCanceled()
+            ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
+            if (!matchesGameType(context, file)) return@p true // check game type at file level
+            consumer.process(file)
+        }
+    }
+
     private fun processResolvedPaths(context: Context, resolved: Set<String>, consumer: Processor<in VirtualFile>): Boolean {
+        // 3.0.2 should be case-sensitive here, should be enough
         if (resolved.isEmpty()) return true
         return FileBasedIndex.getInstance().processFilesContainingAnyKey(ChronicleIndexKeys.FilePath, resolved, context.scope, null, null) p@{ file ->
             ProgressManager.checkCanceled()
@@ -98,10 +137,11 @@ class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSe
     }
 
     private fun processResolvedFileNames(context: Context, resolved: Set<String>, support: ParadoxPathReferenceExpressionSupport, consumer: Processor<in VirtualFile>): Boolean {
+        // 3.0.2 should be case-sensitive here, should be enough
         if (resolved.isEmpty()) return true
         val configExpression = context.configExpression ?: return true
         val resolvedFiles = sortedSetOf<VirtualFile>(compareBy { it.path })
-        FilenameIndex.processFilesByNames(resolved, false, context.scope, null) p@{ file ->
+        FilenameIndex.processFilesByNames(resolved, true, context.scope, null) p@{ file ->
             ProgressManager.checkCanceled()
             val fileInfo = ParadoxAnalysisManager.getFileInfo(file) ?: return@p true // ensure file info is resolved here
             if (!matchesGameType(context, file)) return@p true // check game type at file level
@@ -120,7 +160,6 @@ class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSe
     }
 
     private fun getFilePathsIgnoreLocale(filePath: String): Set<String>? {
-        if (!filePath.endsWith(".yml", true)) return null // 仅限本地化文件
         val configGroup = ChronicleFacade.getConfigGroup()
         val globalLocales = ParadoxLocaleManager.getGlobalLocales(configGroup)
         val localeStrings = globalLocales.mapFast { it.shortId }
@@ -152,12 +191,14 @@ class ParadoxFilePathSearcher : QueryExecutorBase<VirtualFile, ParadoxFilePathSe
 
     fun ParadoxFilePathSearch.Parameters.createContext(): Context {
         val contextElement = selector.file?.toPsiFile(project)
-        return Context(filePath, configExpression, ignoreLocale, contextElement, gameType, project, scope)
+        return Context(filePath, configExpression, ignoreCase, ignoreExtension, ignoreLocale, contextElement, gameType, project, scope)
     }
 
     data class Context(
         val filePath: String?,
         val configExpression: CwtDataExpression?,
+        val ignoreCase: Boolean,
+        val ignoreExtension: Boolean,
         val ignoreLocale: Boolean,
         val contextElement: PsiFile?,
         override val gameType: ParadoxGameType?,